+++ cmd/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.3.3]
+* Add aws log ingest count query config
+* Add azure.RunLogAnalyticsQuery
+* Add mongodb.UpdateAWSMonitoringData
+
 ## [0.3.2]
 * Bugfix: AzureResourceProperties.Other field saving to MongoDB as binary data
 * Add updateResources flag to mongodb.Update command